Fix training scripts for all representations

+++ code/pytorch_train_pos_aclstm.py
@@ -237,7 +237,7 @@ def train(dances, frame_rate, batch, seq_len, read_weight_path, write_weight_fol
             dance=dances[dance_id].copy()
             dance_len = dance.shape[0]
             
-            start_id=random.randint(10, dance_len-seq_len*speed-10)#the first and last several frames are sometimes noisy. 
+            start_id=random.randint(10, int(dance_len - seq_len * speed - 10))#the first and last several frames are sometimes noisy. 
             sample_seq=[]
             for i in range(seq_len):
                 sample_seq=sample_seq+[dance[int(i*speed+start_id)]]
